Move auth from scandium to element-connector.
